feat: partial establishment screen (#2)
* feat: initial screen for estab. users

* feat: added date and time picke

* feat: more create order functions

* feat/fix: fix add to basket listener + estab basket

* feat: added check all items checkbox

* feat: establishment multiple functions + create order

+++ lib/api/product_post_api.dart
@@ -263,7 +263,7 @@ class ProductPost {
       {required User costumer,
       required User publisher,
       required List<Product> products,
-      required int totalPrice}) async {
+      required double totalPrice}) async {
     final orderRef = FirebaseFirestore.instance.collection('orders').doc();
 
     final orders = Orders(
